+++ CHANGELOG.md
@@ -18,8 +18,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix failing `checkoutCustomerAttach` mutation - #9401 by @IKarbowiak
 - Add new mutation `orderCreateFromCheckout` - #9343 by @korycins
 - Add `language_code` field to webhook payload for `Order`, `Checkout` and `Customer` - #9433 by @rafalp
+- Add handling webhook payload via GraphQL subscriptions (#9394)  @jakubkuc
 - Fix access to own resources by App - #9425 by @korycins
-
 
 # 3.1.7